+++ src/db/migrations/0014_product_aisle.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `products` ADD `aisle` text;